# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Go License API
  version: 1.0.0
  description: |
    High-performance multi-tenant license validation service.
    Data Plane under `/licenses/*`, Control Plane under `/admin/*`.
servers:
  - url: http://localhost:8080
    description: Local
paths:
...
  /licenses/validate:
    post:
      summary: Validate a license
      description: Constant-time cache lookup. No DB calls on hot path.
      operationId: postLicenseValidate
      security:
        - TenantAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateRequest'
            examples:
              example:
                value:
                  license_key: "LIC-ABC-123"
                  product_id: "prod_xxxxxxxx"
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateResponse'
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /licenses/{key}/signed:
    get:
      summary: Fetch a signed offline license envelope
      description: Available when signerRegistry is configured. Public cache-only endpoint. Returns a detached-signature JSON envelope (`payload` + `signature`).
      operationId: getSignedLicense
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Signed license envelope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignedLicenseEnvelope'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
    TenantAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
          enum:
            - license_not_active
            - license_expired
            - license_not_found
            - rate_limited
            - unauthorized
            - forbidden
            - validation_failed
        message:
          type: string
      required: [code, message]
      example:
        code: license_expired
        message: License expired on 2026-01-01
    ValidateRequest:
      type: object
      properties:
        license_key:
          type: string
          description: Preferred field; alias 'key' also accepted
        key:
          type: string
          description: Deprecated alias for license_key
        product_id:
          type: string
          description: Preferred field; alias 'product' also accepted
        product:
          type: string
          description: Deprecated alias for product_id
      required: [license_key]
    ValidateResponse:
      type: object
      properties:
        valid:
          type: boolean
        license:
          $ref: '#/components/schemas/ValidationMeta'
        error:
          $ref: '#/components/schemas/Error'
      required: [valid]
      description: On invalid responses, clients should treat `license` as null/absent and use `error` for state.
    ValidationMeta:
      type: object
      properties:
        license_id:
          type: string
        status:
          type: string
        type:
          type: string
          enum: [plan, product]
        plan_id:
          type: string
        plan:
          $ref: '#/components/schemas/ValidationRef'
        product:
          $ref: '#/components/schemas/ValidationRef'
          nullable: true
        product_id:
          type: string
          nullable: true
        not_before:
          type: string
          format: date-time
          nullable: true
          description: License is invalid before this time. Clients SHOULD allow ±5 minutes clock skew.
        expires_at:
          type: string
          format: date-time
          nullable: true
          description: License expiry timestamp. Clients SHOULD allow ±5 minutes clock skew.
        seats_total:
          type: integer
          nullable: true
        unlimited_seats:
          type: boolean
        trial:
          type: boolean
        grace_period_ends_at:
          type: string
          format: date-time
          nullable: true
          description: Set when grace-period information is available. For `license_in_grace_period`, this is non-null.
        features:
          type: array
          items:
            type: string
        version:
          type: integer
        in_grace_period:
          type: boolean
    ValidationRef:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required: [id]
...
    SignedLicenseEnvelope:
      type: object
      properties:
        payload:
          type: string
          description: Canonical JSON string of SignedLicensePayload.
        signature:
          type: string
          description: Base64-encoded Ed25519 signature over raw payload bytes.
        kid:
          type: string
        issuer:
          type: string
      required: [payload, signature, kid, issuer]
...
security: []